requirements.txt
# Python 3.10 Required

# Core Deep Learning (ONNX Runtime for CPU inference)
onnxruntime==1.17.0  # CPU-optimized inference
numpy>=1.23.5,<2.0.0  # Required by ONNX Runtime

# Image Processing (security-patched version)
Pillow==12.0.0

# Web API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.0.0

# Azure Storage (for model loading from blob)
azure-storage-blob==12.18.2
azure-identity>=1.25.1

# Monitoring & Metrics
prometheus-client==0.17.1

# Environment Configuration
python-dotenv==1.0.0

# Security patches (from your Dockerfile)
cryptography>=45.0.7
urllib3>=2.6.3
requests>=2.32.5
starlette>=0.49.1
filelock>=3.20.1


